Make trust check only patch for Tadah 2014
Only the trust check patch is needed for this branch, and I'm also retaining the Output patch for easy debugging in case we need it. Some minor tweaks have been made to the two patches that have been kept to remove anything that's now unneeded and also any tweaks needed to make it work with 2014.